refactor: searching
