+++ gitimoji/Models/FetchState.swift
@@ -1,0 +1,10 @@
+//
+//  FetchState.swift
+//  gitimoji
+//
+//  Created by Timo Zacherl on 23.09.22.
+//
+
+enum FetchState {
+    case loading, success, error, stateless
+}
